Added user handler functions and refactored file and folder names to follow business logic
