+++ app/url_mappings.go
@@ -1,7 +1,12 @@
 package app
 
-import "github.com/jnunortiz/bookstore_users-api/domain/users/controllers"
+import (
+	"github.com/jnunortiz/bookstore_users-api/controllers/ping"
+	"github.com/jnunortiz/bookstore_users-api/controllers/users"
+)
 
 func mapUrls() {
-	router.GET("/ping", controllers.Ping)
+	router.GET("/ping", ping.Ping)
+	router.GET("/users/:user_id", users.GetUser)
+	router.POST("/users", users.CreateUSer)
 }
