+++ controllers/ping/ping_controller.go
@@ -1,4 +1,4 @@
-package controllers
+package ping
 
 import (
 	"net/http"
