+++ controllers/users/user_controller.go
@@ -1,0 +1,15 @@
+package users
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func CreateUSer(c *gin.Context) {
+	c.String(http.StatusNotImplemented, "Implement me!")
+}
+
+func GetUser(c *gin.Context) {
+	c.String(http.StatusNotImplemented, "Implement me!")
+}
