--- domain/users/controllers/user_controller.go
@@ -1,7 +1,0 @@
-package controllers
-
-func CreateUSer() {}
-
-func GetUser() {}
-
-func FindUser() {}
