+++ main.go
@@ -1,8 +1,6 @@
 package main
 
-import (
-	"github.com/jnunortiz/bookstore_users-api/app"
-)
+import "github.com/jnunortiz/bookstore_users-api/app"
 
 func main() {
 	app.StartApplication()
